+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/spel/Typer.scala
@@ -186,17 +186,50 @@ private[spel] class Typer(
       typingResult.map(toNodeResult)
     }
 
+    def typeFieldNameReferenceOnRecord(indexString: String, record: TypedObjectTypingResult): TypingR[TypingResult] = {
+      val fieldIndexedByLiteralStringOpt = record.fields.find(_._1 == indexString)
+      fieldIndexedByLiteralStringOpt.map(f => valid(f._2)).getOrElse {
+        if (dynamicPropertyAccessAllowed) valid(Unknown) else invalid(NoPropertyError(record, indexString))
+      }
+    }
+
+    def typeIndexerOnRecord(indexer: Indexer, record: TypedObjectTypingResult) = {
+      withTypedChildren {
+        case TypedObjectWithValue(_, indexString: String) :: Nil =>
+          // Children are typed in a non-obvious way in case of PropertyOrFieldReference
+          indexer.children match {
+            case (ref: PropertyOrFieldReference) :: Nil => typeFieldNameReferenceOnRecord(ref.getName, record)
+            case _                                      => typeFieldNameReferenceOnRecord(indexString, record)
+          }
+        case indexKey :: Nil if indexKey.canBeSubclassOf(Typed[String]) =>
+          if (dynamicPropertyAccessAllowed) valid(Unknown) else invalid(DynamicPropertyAccessError)
+        case _ :: Nil =>
+          indexer.children match {
+            case (ref: PropertyOrFieldReference) :: Nil => typeFieldNameReferenceOnRecord(ref.getName, record)
+            case _ => if (dynamicPropertyAccessAllowed) valid(Unknown) else invalid(DynamicPropertyAccessError)
+          }
+        case _ =>
+          invalid(IllegalIndexingOperation)
+      }
+    }
+
     @tailrec
     def typeIndexer(e: Indexer, typingResult: TypingResult): NodeTypingResult = {
       typingResult match {
         case TypedClass(clazz, param :: Nil)
             if clazz.isAssignableFrom(classOf[java.util.List[_]]) || clazz.isAssignableFrom(classOf[Array[Object]]) =>
+          // TODO: validate indexer key - the only valid key is an integer - but its more complicated with references
           validNodeResult(param)
         case TypedClass(clazz, keyParam :: valueParam :: Nil) if clazz.isAssignableFrom(classOf[java.util.Map[_, _]]) =>
           validNodeResult(valueParam)
         case d: TypedDict                    => dictTyper.typeDictValue(d, e).map(toNodeResult)
         case TypedUnion(possibleTypes)       => typeUnion(e, possibleTypes)
         case TypedTaggedValue(underlying, _) => typeIndexer(e, underlying)
+        case r: TypedObjectTypingResult      => typeIndexerOnRecord(e, r)
+        // TODO: add indexing on strings
+        // TODO: how to handle other cases?
+        case TypedNull =>
+          invalidNodeResult(IllegalIndexingOperation)
         case _ =>
           val w = validNodeResult(Unknown)
           if (dynamicPropertyAccessAllowed) w else w.tell(List(DynamicPropertyAccessError))