+++ interpreter/src/test/scala/pl/touk/nussknacker/engine/spel/TyperSpec.scala
@@ -5,18 +5,33 @@ import cats.data.ValidatedNel
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import org.springframework.expression.common.TemplateParserContext
+import org.springframework.expression.spel.standard
 import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.generics.ExpressionParseError
 import pl.touk.nussknacker.engine.api.typed.supertype.{CommonSupertypeFinder, SupertypeClassResolutionStrategy}
-import pl.touk.nussknacker.engine.api.typed.typing.Typed
 import pl.touk.nussknacker.engine.definition.clazz.ClassDefinitionSet
+import pl.touk.nussknacker.engine.api.typed.typing.{
+  Typed,
+  TypedNull,
+  TypedObjectTypingResult,
+  TypedObjectWithValue,
+  Unknown
+}
 import pl.touk.nussknacker.engine.dict.{KeysDictTyper, SimpleDictRegistry}
 import pl.touk.nussknacker.engine.expression.PositionRange
+import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.IllegalOperationError.DynamicPropertyAccessError
+import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.MissingObjectError.NoPropertyError
 import pl.touk.nussknacker.engine.spel.Typer.TypingResultWithContext
+import pl.touk.nussknacker.engine.spel.TyperSpecTestData.TestRecord._
 import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
 import pl.touk.nussknacker.test.ValidatedValuesDetailedMessage
 
 class TyperSpec extends AnyFunSuite with Matchers with ValidatedValuesDetailedMessage {
+
+  private implicit val defaultTyper: Typer = buildTyper()
+  private val dynamicAccessTyper: Typer    = buildTyper(dynamicPropertyAccessAllowed = true)
+  private val parser: standard.SpelExpressionParser =
+    new org.springframework.expression.spel.standard.SpelExpressionParser()
 
   test("simple expression") {
     typeExpression("#x + 2", "x" -> 2) shouldBe Valid(
@@ -65,31 +80,93 @@ class TyperSpec extends AnyFunSuite with Matchers with ValidatedValuesDetailedMe
       s"Cannot do projection/selection on ${Typed.fromInstance(1).display}"
   }
 
-  private val strictTypeChecking               = false
-  private val strictMethodsChecking            = false
-  private val staticMethodInvocationsChecking  = false
-  private val methodExecutionForUnknownAllowed = false
-  private val dynamicPropertyAccessAllowed     = false
-  private val classResolutionStrategy          = SupertypeClassResolutionStrategy.Union
-  private val commonSupertypeFinder            = new CommonSupertypeFinder(classResolutionStrategy, strictTypeChecking)
-  private val dict                             = new SimpleDictRegistry(Map.empty)
+  test("indexing on records for primitive types") {
+    typeExpression(s"$testRecordExpr['string']").validValue.finalResult.typingResult shouldBe
+      TypedObjectWithValue(Typed.typedClass[String], "stringVal")
+    typeExpression(s"$testRecordExpr['int']").validValue.finalResult.typingResult shouldBe
+      TypedObjectWithValue(Typed.typedClass[Int], 1)
+    typeExpression(s"$testRecordExpr['boolean']").validValue.finalResult.typingResult shouldBe
+      TypedObjectWithValue(Typed.typedClass[Boolean], true)
+    typeExpression(s"$testRecordExpr['null']").validValue.finalResult.typingResult shouldBe
+      TypedNull
+  }
 
-  private val typer = new Typer(
-    commonSupertypeFinder,
-    new KeysDictTyper(dict),
-    strictMethodsChecking,
-    staticMethodInvocationsChecking,
-    ClassDefinitionSet.forDefaultAdditionalClasses,
+  test("indexing on records with string literal index") {
+    typeExpression(s"$testRecordExpr['string']").validValue.finalResult.typingResult shouldBe
+      TypedObjectWithValue(Typed.typedClass[String], "stringVal")
+  }
+
+  test("indexing on records with variable reference index") {
+    typeExpression(s"$testRecordExpr[#var]", "var" -> "string").validValue.finalResult.typingResult shouldBe
+      TypedObjectWithValue(Typed.typedClass[String], "stringVal")
+  }
+
+  test("indexing on records with property reference index") {
+    typeExpression(s"$testRecordExpr[string]").validValue.finalResult.typingResult shouldBe
+      TypedObjectWithValue(Typed.typedClass[String], "stringVal")
+  }
+
+  test("indexing on records for record values") {
+    typeExpression(s"$testRecordExpr['nestedRecord']").validValue.finalResult.typingResult shouldBe
+      TypedObjectTypingResult(Map("nestedRecordKey" -> TypedObjectWithValue(Typed.typedClass[Int], 2)))
+  }
+
+  test("indexing on records for nested record values") {
+    typeExpression(
+      s"$testRecordExpr['nestedRecord']['nestedRecordKey']"
+    ).validValue.finalResult.typingResult shouldBe
+      TypedObjectWithValue(Typed.typedClass[Int], 2)
+  }
+
+  test(
+    "indexing on records with string literal or variable reference for not present keys returns error when dynamic access is disabled"
+  ) {
+    typeExpression(s"$testRecordExpr['$nonPresentKey']").invalidValue.toList should matchPattern {
+      case NoPropertyError(typingResult, key) :: Nil if typingResult == testRecordTyped && key == nonPresentKey =>
+    }
+    typeExpression(s"$testRecordExpr[#var]", "var" -> s"$nonPresentKey").invalidValue.toList should matchPattern {
+      case NoPropertyError(typingResult, key) :: Nil if typingResult == testRecordTyped && key == nonPresentKey =>
+    }
+    // TODO: this behavior is to be fixed - ideally this should behave the same as above
+    typeExpression(s"$testRecordExpr[$nonPresentKey]").invalidValue.toList should matchPattern {
+      case NoPropertyError(typingResult, key) :: DynamicPropertyAccessError :: Nil
+          if typingResult == testRecordTyped && key == nonPresentKey =>
+    }
+  }
+
+  test("indexing on records for not present keys returns unknown type or error when dynamic access is enabled") {
+    typeExpression(s"$testRecordExpr['$nonPresentKey']")(
+      dynamicAccessTyper
+    ).validValue.finalResult.typingResult shouldBe
+      Unknown
+    typeExpression(s"$testRecordExpr[#var]", "var" -> nonPresentKey)(
+      dynamicAccessTyper
+    ).validValue.finalResult.typingResult shouldBe
+      Unknown
+    typeExpression(s"$testRecordExpr[$nonPresentKey]")(dynamicAccessTyper).invalidValue.toList should matchPattern {
+      case NoPropertyError(typingResult, key) :: Nil if typingResult == testRecordTyped && key == nonPresentKey =>
+    }
+  }
+
+  private def buildTyper(dynamicPropertyAccessAllowed: Boolean = false) = new Typer(
+    commonSupertypeFinder = new CommonSupertypeFinder(
+      classResolutionStrategy = SupertypeClassResolutionStrategy.Union,
+      strictTaggedTypesChecking = false
+    ),
+    dictTyper = new KeysDictTyper(new SimpleDictRegistry(Map.empty)),
+    strictMethodsChecking = false,
+    staticMethodInvocationsChecking = false,
+    classDefinitionSet = ClassDefinitionSet.forDefaultAdditionalClasses,
     evaluationContextPreparer = null,
-    methodExecutionForUnknownAllowed,
-    dynamicPropertyAccessAllowed
+    methodExecutionForUnknownAllowed = false,
+    dynamicPropertyAccessAllowed = dynamicPropertyAccessAllowed
   )
-
-  private val parser = new org.springframework.expression.spel.standard.SpelExpressionParser()
 
   private def typeExpression(
       expr: String,
       variables: (String, Any)*
+  )(
+      implicit typer: Typer
   ): ValidatedNel[ExpressionParseError, CollectedTypingResult] = {
     val parsed        = parser.parseExpression(expr)
     val validationCtx = ValidationContext(variables.toMap.mapValuesNow(Typed.fromInstance))
@@ -99,10 +176,37 @@ class TyperSpec extends AnyFunSuite with Matchers with ValidatedValuesDetailedMe
   private def typeTemplate(
       expr: String,
       variables: (String, Any)*
+  )(
+      implicit typer: Typer
   ): ValidatedNel[ExpressionParseError, CollectedTypingResult] = {
     val parsed        = parser.parseExpression(expr, new TemplateParserContext())
     val validationCtx = ValidationContext(variables.toMap.mapValuesNow(Typed.fromInstance))
     typer.typeExpression(parsed, validationCtx)
+  }
+
+}
+
+object TyperSpecTestData {
+
+  object TestRecord {
+    val nonPresentKey: String = "nonPresentKey"
+    val testRecordExpr: String =
+      "{int: 1, string: 'stringVal', boolean: true, 'null': null, nestedRecord: {nestedRecordKey: 2}}"
+
+    val testRecordTyped: TypedObjectTypingResult = TypedObjectTypingResult(
+      Map(
+        "string"  -> TypedObjectWithValue(Typed.typedClass[String], "stringVal"),
+        "null"    -> TypedNull,
+        "boolean" -> TypedObjectWithValue(Typed.typedClass[Boolean], true),
+        "int"     -> TypedObjectWithValue(Typed.typedClass[Int], 1),
+        "nestedRecord" -> TypedObjectTypingResult(
+          Map(
+            "nestedRecordKey" -> TypedObjectWithValue(Typed.typedClass[Int], 2)
+          )
+        ),
+      )
+    )
+
   }
 
 }